+++ resources/views/layouts/app.blade.php
@@ -5,14 +5,18 @@
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
         <meta name="csrf-token" content="{{ csrf_token() }}" />
         <title>ErrorLens - {{ config('app.name') }}</title>
+
+        <link rel="icon" type="image/x-icon" href="{{ asset('vendor/error-lens/assets/img/favicon.png') }}" />
         
         <link rel="stylesheet" href="{{ asset('vendor/error-lens/assets/css/bootstrap.min.css') }}" />
         <link rel="stylesheet" href="{{ asset('vendor/error-lens/assets/css/style.css') }}" />
     </head>
     <body>
-        <nav class="navbar navbar-expand-lg navbar-dark bg-dark fixed-top">
+        <nav class="navbar navbar-expand-lg navbar-dark main_navbar">
             <div class="container">
-                <a class="navbar-brand" href="{{ route('error-lens.index') }}">ErrorLens</a>
+                <a class="navbar-brand" href="{{ route('error-lens.index') }}">
+                    <img src="{{ asset('vendor/error-lens/assets/img/error-lens.png') }}" alt="" style="width: 180px;" />
+                </a>
                 
                 <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
@@ -23,14 +27,15 @@
                         <li class="nav-item">
                             <form action="{{ route('error-lens.clear') }}" method="post">
                                 @csrf
-                                <a class="nav-link" id="clear-logs" href="{{ route('error-lens.clear') }}">Clear Logs</a>
+                                <a class="nav-link text-white" id="clear-logs" href="{{ route('error-lens.clear') }}">Clear Logs</a>
                             </form>
                         </li>
                     </ul>
                 </div>
             </div>
         </nav>
-        <div class="container-fluid mt-5">
+        
+        <div class="container-fluid px-4">
             @yield('content')
         </div>