+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# Changelog
+# Release Notes
 
-All notable changes to `error-lens` will be documented in this file.
+## [v1.x-beta](https://github.com/narolalabs/error-lens/compare/main...v1.x-beta) - 2023-09-14
+
+### Added
+
+- Added support for Laravel ^8.x and PHP ^7.3 versions by @narolalebs
+- Improve display error logs that help to find out actual errors by @narolalebs
+- Improve UI/UX by @narolalebs